CREATE: product file
