+++ src/controllers/index.js
@@ -1,3 +1,3 @@
-const productController = require("./productController");
+const productController = require("./product.controller");
 
 module.exports = { productController };
