+++ src/controllers/product.controller.js
@@ -2,8 +2,16 @@ const { productService } = require("../services");
 
 const getProduct = async (req, res) => {
   try {
+    // const { page, sort, category } = req.query;
+    // console.log(queryParams);
+    // console.log(Object.entries(queryParams));
     const userId = req.user_id;
-    const productList = await productService.getProductList(userId);
+    const productList = await productService.getProductList(
+      userId
+      // page,
+      // sort,
+      // category
+    );
     const totalProduct = await productService.getTotalProduct();
     res.status(200).json({
       message: "querySuccess",
