+++ src/services/index.js
@@ -1,3 +1,3 @@
-const productService = require("./productService");
+const productService = require("./product.service");
 
 module.exports = { productService };
